UPSTREAM: intel/minnowmax: Program GPIO for power LED
MinnowBoard Turbot systems have a GPIO-controlled LED that is
generally used to indicate the CPU is running.  Commit 2ae9cce8
changed the parameter for GPIO_NC, exposing an issue with the
assumed behavior of the signal.  Use a pull-down to turn on the
LED.

BUG=None
BRANCH=None
TEST=None

Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Reviewed-on: https://review.coreboot.org/16428
Tested-by: build bot (Jenkins)
Reviewed-by: Martin Roth <martinroth@google.com>

Change-Id: I153870904c007d89016c0d47bb3db9b824ebbcff
Reviewed-on: https://chromium-review.googlesource.com/381668
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>