UPSTREAM: intel/gma: Fix typo GMBUS0 -> GMBUS1 in edid.c
This typo existed in code before rewriting for using
defines and it's clearly visible after rewrite.
Previously it was writing to reserved area of GMBUS0 register,
while values are matching those of GMBUS1.

This line probably is a no-op since it's just sending the STOP
again (without an address set this time).

BUG=none
BRANCH=none
TEST=none

Change-Id: I7bcbaff545f45f0bcb6c23d7f4496f10681ef2eb
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: 34e10871f9f1004a0c007a26d2d77440997fafa1
Original-Change-Id: Ic85ef925c41ad01ed469f9d4f4412cbe44ca6d8e
Original-Signed-off-by: Sebastian "Swift Geek" Grzywna <swiftgeek@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/16341
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/524605